Implement significant speed-up
- Before this change, the code would calculate all transitive 
dependencies for a direct dependency (e.g. `maybe`) each time it 
appeared as a transitive dependency, no matter how deep within the 
transitive dependency graph it was. Thus, if Package A depended on B, C, 
and D, and B, C, and D all depended on `maybe`, we would calculate 
`maybe`'s transitive dependencies three times.
- After this change, we only calculate `maybe`'s transitive dependencies 
once. 
- Moreover, there's another speed up. If Package A depends on package B, 
and we already know package B's dependencies, we immediately return with 
package B's dependencies. In other words, we build the dependency graph 
from the bottom up (or near that)